+++ app/src/main/java/com/darshan/notificity/NotificationsActivity.kt
@@ -35,6 +35,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.darshan.notificity.ui.theme.NotificityTheme
+import com.darshan.notificity.utils.Util
 
 class NotificationsActivity : ComponentActivity() {
     private val repository: NotificationRepository by lazy { NotificationRepository(AppDatabase.getInstance(application).notificationDao()) }
