+++ app/src/main/java/com/darshan/notificity/utils/Util.kt
@@ -1,4 +1,4 @@
-package com.darshan.notificity
+package com.darshan.notificity.utils
 
 import java.time.Instant
 import java.time.ZoneId
